openpilot v0.7 release

+++ launch_chffrplus.sh
@@ -14,12 +14,15 @@ function launch {
   # Wifi scan
   wpa_cli IFNAME=wlan0 SCAN
 
-  # apply update
-  if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
-    git reset --hard @{u} &&
-    git clean -xdf &&
-
-    exec "${BASH_SOURCE[0]}"
+  # apply update only if no_ota_updates does not exist in /data directory
+  file="/data/no_ota_updates"
+  if ! [ -f "$file" ]; then
+    if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
+      git reset --hard @{u} &&
+      git clean -xdf &&
+      
+      exec "${BASH_SOURCE[0]}"
+    fi
   fi
 
   # no cpu rationing for now
